Allow switching to business context.
- Recycle classes from previous abandoned implementation.
