+++ templates/_partials/user/account_dropdown.html.twig
@@ -9,18 +9,18 @@
       </a>
     </li>
 
-    {% if is_granted('ROLE_PRO') %}
-      {% set queryParams = app.request.query.all %}
-      {% if context_channel.channel.code == 'pro' %}
-        {% set queryParams = queryParams|merge({change_channel: 'web'}) %}
+    {% if is_granted('ROLE_BUSINESS') %}
+      {% set query_params = app.request.query.all %}
+      {% if business_context.active %}
+        {% set query_params = query_params|merge({ _business: 'false' }) %}
         {% set accountType = 'navbar.lambdaCart' %}
       {% else %}
-        {% set queryParams = queryParams|merge({change_channel: 'pro'}) %}
+        {% set query_params = query_params|merge({ _business: 'true' }) %}
         {% set accountType = 'navbar.proCart' %}
       {% endif %}
-      {% set currentPath = path(app.request.attributes.get('_route'),
-        app.request.attributes.get('_route_params')|merge(queryParams)) %}
-      <li><a href="{{ currentPath }}">{{ accountType|trans }}</a></li>
+      {% set current_path = path(app.request.attributes.get('_route'),
+        app.request.attributes.get('_route_params')|merge(query_params)) %}
+      <li><a href="{{ current_path }}">{{ accountType|trans }}</a></li>
     {% endif %}
 
     <li role="separator" class="divider"></li>
